Add {partialLDSC}
This is an extension of LDSC / GenomicSEM with the key extension being including a block-jackknife approach to assess whether differences between the estimates (genetic correlation before/after adjustment for the genetics of a confounder) were statistically meaningful (developer = Ninon Mounier https://github.com/n-mounier). There is a branch in development to do this for multiple confounders simultaneously.